rand_distr: Test normal distribution
`Normal` is sampled many times into a histogram, which is then compared
to the expected probability density function. To make debugging easier,
sparklines of the expected distribution, the histogram, and their
difference are printed.

Currently, the test fails if the difference is significantly larger than
the expected error of the histogram bin. However, the error estimate
does not take the error in the normalization due to the finite width of
the histogram into account. This should not be a problem, as long as the
distribution is almost zero outside the range covered by the histogram.

In principle, this approach can be generalized to other distributions.