Fix: Added missing import for `no_std` build.
Before this commit, the command
```bash
rustup run "1.63.0" cargo build --no-default-features
```
failed in the line
```rust
let test_x = if symmetric { x.abs() } else { x };
```
in file `src/utils.rs` with the error that `abs()` is not a method of `f64`. This is true for all `rustc` version up to `1.83.0`. From `1.84.0` onwards, this appears to be fixed without the change. However, the change is still necessary for `no_std` builds on older versions of `rustc`.