Avoid returning NaN from Gamma::sample
This changes the order of multiplications used to compute the result
to avoid multiplying zero with an expression that can overflow to +inf.

Note that the parameter combinations which could lead to this (shape
very close to zero, scale very close to the max float value) continue to
be inaccurately handled; the Gamma distribution sampler will now tend to
return zero instead of NaN for them. The limit (shape 0, scale inf) is
not well defined.